refactor: подключен root-reducer и удален reducer
