+++ src/components/blocks/movie-card-tabs/movie-card-tab-reviews/movie-card-tab-reviews.jsx
@@ -7,8 +7,7 @@ import {fetchReviews} from "../../../../store/movie-data/movie-data-api-actions"
 
 const MovieCardTabReviews = ({id}) => {
 
-  const reviews = useSelector((state) => state.reviews);
-  const isReviewsLoaded = useSelector((state) => state.isReviewsLoaded);
+  const {reviews, isReviewsLoaded } = useSelector((state) => state.DATA_ITEM);
   const dispatch = useDispatch();
 
   const reviewsItems = reviews.length !== 0 ? getTwoArraysFromOne(reviews) : reviews;
