+++ src/store/error/error-reducer.js
@@ -1,4 +1,4 @@
-import {errorAction} from "./error-action";
+import {catchError} from "./error-action";
 import {createReducer} from "@reduxjs/toolkit";
 
 const initialState = {
@@ -6,7 +6,7 @@ const initialState = {
 };
 
 export const errorReducer = createReducer(initialState, (builder) => {
-  builder.addCase(errorAction, (state) => {
+  builder.addCase(catchError, (state) => {
     state.isCatchError = true;
   });
 });
