+++ src/store/movies-filter/movies-filter-reducer.js
@@ -1,4 +1,4 @@
-import {MOVIES_COUNT_PER_STEP} from "../constants/common";
+import {MOVIES_COUNT_PER_STEP} from "../../constants/common";
 import {changeGenre, resetFilter, showMoreMovies} from "./movies-filter-action";
 import {createReducer} from "@reduxjs/toolkit";
